+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## --WIP--
 
+### Additions
+
+* Added the transmission library tab
+  * You can edit, preview, save, load, and copy transmissions
+
 ### Changes
 
 * Editing an entry you don't have defined now auto-selects the name
@@ -10,6 +15,7 @@
 * Theme colors are less washed out
 * Changed font
 * Removed button click noises
+* Reaching the signal cap now displays how many signals are in the message, up to 4096
 
 ### Fixes